requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
pytesseract==0.3.10
Pillow==10.0.0
PyMuPDF==1.23.8
reportlab==4.0.4
werkzeug==2.3.7